split Store.ts
